src/maintenance/requirements.txt
azure-functions
requests
cryptography
azure-mgmt-cosmosdb
azure-keyvault
azure-identity
azure-mgmt-web